refactor: Replace f64 with Decimal for all financial calculations
- Update Trade struct to use Decimal for all financial fields (amount_sol, pnl_sol, pnl_usd, costs)
- Update SignalPayload struct to use Decimal for amount_sol
- Update ProfitTargetState struct to use Decimal for all price/amount fields
- Update database functions to accept Decimal and convert to f64 at SQLite boundary
- Update all PnL calculations to use Decimal throughout (profit_targets, momentum_exit, executor)
- Add rust_decimal serde feature for API serialization
- Maintain backward compatibility with database (REAL type) and external APIs

This eliminates IEEE 754 floating point errors in currency calculations (SOL, USD)
as recommended in the code review. All financial calculations now use precise
decimal arithmetic via rust_decimal::Decimal.

+++ operator/src/models/trade.rs
@@ -3,6 +3,7 @@
 use super::{Action, Signal, Strategy};
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
+use rust_decimal::Decimal;
 
 /// Trade status representing the state machine
 ///
@@ -132,7 +133,7 @@ pub struct Trade {
     /// Trade action
     pub side: Action,
     /// Amount in SOL
-    pub amount_sol: f64,
+    pub amount_sol: Decimal,
     /// Current status
     pub status: TradeStatus,
     /// Number of retry attempts
@@ -142,19 +143,19 @@ pub struct Trade {
     /// Error message (if failed)
     pub error_message: Option<String>,
     /// Realized PnL in SOL
-    pub pnl_sol: Option<f64>,
+    pub pnl_sol: Option<Decimal>,
     /// Realized PnL in USD
-    pub pnl_usd: Option<f64>,
+    pub pnl_usd: Option<Decimal>,
     /// Jito tip paid (in SOL)
-    pub jito_tip_sol: Option<f64>,
+    pub jito_tip_sol: Option<Decimal>,
     /// DEX fee paid (in SOL)
-    pub dex_fee_sol: Option<f64>,
+    pub dex_fee_sol: Option<Decimal>,
     /// Slippage cost (in SOL)
-    pub slippage_cost_sol: Option<f64>,
+    pub slippage_cost_sol: Option<Decimal>,
     /// Total cost (Jito tip + DEX fee + slippage) in SOL
-    pub total_cost_sol: Option<f64>,
+    pub total_cost_sol: Option<Decimal>,
     /// Net PnL after all costs (in SOL)
-    pub net_pnl_sol: Option<f64>,
+    pub net_pnl_sol: Option<Decimal>,
     /// Creation timestamp
     pub created_at: DateTime<Utc>,
     /// Last update timestamp
